Fix build for darwin and windows
