+++ pkg/collect/host_filesystem_performance_windows.go
@@ -1,0 +1,7 @@
+package collect
+
+import "github.com/pkg/errors"
+
+func HostFilesystemPerformance(c *HostCollector) (map[string][]byte, error) {
+	return nil, errors.New("Filesystem performance collector is only implemented for Linux")
+}
